+++ next.config.js
@@ -4,13 +4,13 @@ const nextConfig = {
   webpack(config) {
     config.resolve.modules.push(__dirname); // 추가
     return config;
-  }
-}
+  },
+};
 module.exports = {
   images: {
-    disableStaticImages: true
-  }
-}
+    disableStaticImages: true,
+  },
+};
 const withImages = require('next-images');
 module.exports = withImages;
 module.exports = nextConfig;
