+++ src/containers/mapping.tsx
@@ -8,3 +8,5 @@
 // }
 //
 // export default Mapping;
+
+export {}
