--- src/pages/Home/Style.ts
@@ -1,10 +1,0 @@
-import styled from "@emotion/styled";
-
-export const Positioner = styled.div`
-    width: 100%;
-    padding: 15vh 0 15vh 15vh;
-`;
-
-export const Components = styled.div`
-    display: flex;
-`;
