--- src/pages/write/Style.ts
@@ -1,7 +1,0 @@
-import styled from "@emotion/styled";
-
-export const Positioner = styled.div`
-    display: flex;
-    width: 100%;
-    padding: 10vh 15vh 5vh 15vh;
-`;
