--- src/pages/write/index.tsx
@@ -1,12 +1,0 @@
-import * as S from './Style';
-import { Write } from '../../components/index';
-
-const WritePage = () => {
-    return(
-        <S.Positioner>
-            <Write />
-        </S.Positioner>
-    );
-}
-
-export default WritePage;
